refactor(ui): split upgrade-modal variant content into separate files
UpgradeModalContent.tsx exceeded the 300-line SRP limit by bundling the
LimitContent, SeatsContent, and LiteMemberRestrictionContent variants
with the shared MODAL_CONTENT map and UpgradeModal shell. Each variant
now lives in its own file under upgrade-modal/, matching the codebase's
per-concern directory convention. Public API (UpgradeModal, MODAL_CONTENT)
and the lazy-loaded import path are unchanged.
